SERVER-3991 (reissue) -- don't index past end of string
This is a reissue of my earlier fix. I'm trying to break them down into
bite-sized parts to make them easier to pull.

Break out of the for() loop once we've advanced the counter past the end.
Skip the test for dangling operator (no character to examine, we're inside
a quoted string) and set danglingOp = false.


Former-commit-id: 443ecbfb44e7a474749f3f40f4f877aa41b28e6e