SERVER-4538  detect if pages are in our working set, on Windows
If the page is not in our working set, this function will return false.
This will be true for both hard and soft pagefaults; however, it is expected
under normal circumstances that most pagefaults will be hard (need to be loaded
from the pagefile on disk into RAM).
On a server with nothing else on it, the OS ought to allow mongod's working set
to expand to "most" of RAM, thus avoiding many soft pagefaults.


Former-commit-id: 15d5ed29163283a784fab344f50404a0a46207af