turn on boost asserts for all compilation
The legacy linux build has been failing in release mode because disabling boost asserts results in an unused variable which triggers a warning, and we have warnings-as-errors turned on in that build.
To avoid the warning, and to improve code in general, I have turned on boost asserts for all compilation.  This means if you hit a boost assert, it will call mongo::log() with the message. In debug mode,
the program will subsequently abort.


Former-commit-id: 30163a270db97c479a843a8c3e8ac1b644ad759f