SERVER-4312 Fix two display bugs in linenoise
I broke the fetching of the screen width with some code rearrangement
weeks ago and finally dug into it.  Can't use console_out to fetch
screen width because it is set in enableRawMode() which is (now)
called later.  This caused bizarre behavior where the cursor would
walk up the screen as you recalled previous history ... I thought this
was UTF-8 related, it isn't.  Second bug is trying to refresh the line
after displaying "Display all %d possibilities?" in tab completion:
don't do it in that case, we already fixed the line and moved to the
next one.


Former-commit-id: f84147cf3fa29aea65529ac5cce3c56678530a48