SERVER-4312 Don't read kill buffer that isn't there
If your first "kill" operation is for zero characters, the initial
kill ring buffer isn't saved.  If this is immediately followed by
a kill operation that does kill text, the code was trying to join
the two kills, but the first one never got a buffer.  Only join
text when there is something to join ... if no buffer to join with,
just store the new text in a slot by itself.  Most visible in debug
builds, but a bug in any build.
Steps to reproduce:
1) Start shell
2) Type "a"
3) Press Ctrl-K
4) Press Ctrl-U


Former-commit-id: 3a59385162569f7d2dbf0aec0bfb03c52ace706c