Roll back most of the following two patches:
39ac424b8134b62b85805f3eab9709efa65f6fab [formerly c4851043c682c9e2a9230166045ce85b518b85eb]:  qlock can ask if in greedy mode or not
48a7f34bdafafed578c9c85a83392e30507880b9 [formerly 938ac1619dc88dc5f9e6df47dcb981fbf0f33663]: if in ThreadSpanningOp don't profile

The problem is that isGreedSuspended cannot be called with the qlock's internal
mutex already held.  However, a caller not holding that lock already can't
know that by the time the function returns, its return value is still valid.
Another thread may change it in parallel.  The result is that the attempt
to use it to stop profiling when thread-spanning ops are enabled only appears
to provide that protection.


Former-commit-id: b8cec6e3c0df43c63e096dabd93d5697c6143b82