SERVER-5805 -- Fix error uncovered by dur_passthrough fsync.js.
When a Lock::GlobalWrite goes out of scope, it must uncondtionally undo any greediness changes made
when the Lock::GlobalWrite was constructed.  If the object stopped greed at construction, was subsequently
downgraded to global read state ('R'), and then went out of scope, it must do extra work to ensure that
greed is restarted.  This change reintroduces that extra work, which was erroneously removed in an earlier
patch.


Former-commit-id: e7e665ca44e2b93bb08b78d0e116e4ebf5bb2c41