SERVER-6960 Always link object files and static archives in the same order.
SCons (correctly) considers two link command lines different if the order of
object files or static libraries changes.  As a result, the libdeps system needs
to produce consistent ordering of these files in the $_LIBDEPS expansion.  This
patch achieves this by sorting the _LIBDEPS candidate expansion by the string
name of the expanded objects.

This problem is akin to that from SERVER-5254, which was solved in a similar
manner.


Former-commit-id: 645aa497732e829fd43092f02102ac37cc3a8b08