SERVER-6671 suppress unique index enforcement during STARTUP2 or RECOVERING secondary states
Operations involving unique indexes may be repeated during a recovery or initial sync. We need to ignore unique key violations in this situation.  This is okay because by the time we reach the end of the oplog stream (and are finally caught up), all the violations should be resolved (as we should end up with an exact copy of the primary).


Former-commit-id: 1e852c70e5ad7907ec97142d27278a21324ae544