SERVER-7410 Fix SYSLIBDEPS functionality
The previous implementation was not recursively following the LIBDEPS hierarchy, looking for SYSLIBDEPS
to add.  Instead, it was trying to walk a hierachy like LIBDEPS rooted in the SYSLIBDEPS variable.  This
isn't really a meaningful behavior, since SYSLIBDEPS always list system libraries, which are always leaf
nodes in the dependency graph for our build system.


Former-commit-id: fdebd7bae6ca938a80713694ef2dfe360f3dc27f