SERVER-9621 Disable jstests/dur/closeall.js on 32-bit Windows
This test fails frequently on 32-bit Windows due to remapPrivateView
being unable to remap a file when another thread allocate memory
while a file is unmapped.

Disable the test in 32-bit Windows since this can't really be fixed
without changes to the storage layer (and 32-bit does not benefit
from stress testing).


Former-commit-id: 33e52874dfcfd2257f49bb1b4007664723d7d309