SERVER-9824 Retry hash table grow operation on excessive collisions
Be defensive against excessively long hash chains in a resized table in
UnorderedFastKeyTable::_grow().  Retry growing the table up to 5 times
before declaring failure (as we do when inserting into a table).  Reduce
the maximum retry count for table insertions from 10 to 5, since growing
the table by a factor of 32 (2**5) guarantees that no hash chain can be
as long as 5% (1/20, the default _maxProbeRatio) of the new table size.

Former-commit-id: df2274d2a2cac0fe41de61b0a7943ff7977975f7