SERVER-7404 Link to posix_fadvise at runtime for Solaris
For the Solaris/SmartOS build, do not make direct calls to posix_fadvise,
which is present in Solaris 11 but not in Solaris 10.  Instead, see if
it is available in a loaded library (which will be libc.so.1) at runtime
and either call it or call an emulation.  The emulation is a no-op.


Former-commit-id: 69d7632211e20d05f564e67e0cb9fd98929e5d5a