SERVER-14624 Local validation methods for replica set reconfiguration.
The introduced functions, repl::validateConfigForInitiate() and
repl::validateConfigForReconfig(), confirm that it is OK to initiate/reconfigure
a replica set with a given new configuration, assuming that the reconfiguration
is launched from a particular (primary) node.  Unlike the validate method on
ReplicaSetConfig, these functions consider information about the current node
and the prior replica set configuration.

This implementation absorbs the local check functionality of
checkMembersUpForConfigChange(), but does not incorporate the checks from
ReplicaSetConfig::legalChange.  The latter checks will be added in another
patch.

Former-commit-id: 9130687a561a3020ba3f1c408fb2f6d47af46a71