SERVER-14219 logOp implicit collection creation
This patch only logs the creation of collections that will not be deleted on
operation failure. This fixes a bug where the collection would be created on
master, but if the operation failed, the slaves would not get the logOp that
implicitly created the collection leading them to differ from the master.


Former-commit-id: 081f11c248433f74806b7508555b212bc4ff0164