SERVER-15111 Treat corruption of final journal section as an expected event
This was always the intention of the code, as a crash may happen while
writing the latest journal section. The code wasn't adjusted when we started
preallocating and reusing journal files and when we started compressing
journal sections, so it ended up treating many common types of corruption as
errors.


Former-commit-id: 77b11e686c873ea44e14e6f7419adb24d5f40106