SERVER-14561 Heartbeat callback and scheduling.
With this change, the ReplicationCoordinatorImpl should have full heartbeat support, including
primary step-down, standing for election, and heartbeat reconfig.  Unit test coverage is light,
but present.  Every test that needs to set member state to primary or fall back to secondary
exercises the code, along with a unit test of the joining-a-new-node-to-a-replica-set behavior.


Former-commit-id: 5f0918f1e83eb48f5855df806f7b911029a6aa5c