SERVER-15537 Make methods called by bgsync and fetch threads robust to node being removed from set.
If a node is removed from a replica set, or if it is not a member of its locally
stored configuration, the ReplicationCoordinator must still respond non-fatally
to method calls from the bgsync and fetch threads.  It's typically OK to throw
an exception or return an error value, depending on circumstance.  In this
change, we make one of each change.


Former-commit-id: 863423d8cfa45a0420d8139a808cb61536a77d8b