SERVER-15549 listDatabases command should only list existent databases
The database information is split between the storage-engine database
catalog and the database metadata, which could potentially be in-memory,
but not created on disk yet. The cases where this happens are only
applicable to the Admin database.

This change makes the listDatabases command only list databases from the
storage-engine's catalog.


Former-commit-id: c46e86610205985ae9744685e62fe269e03835e1