SERVER-15535 When opting to sleep before standing for election, schedule a wake-up event.
The prior behavior in the new replication implementation was to wait for the
next incoming heartbeat.  However, if heartbeats are synchronized, this defeats
the purpose of the random sleep, which is to prevent two nodes from repeatedly
standing for election at the same time.  This is a problem because the sleep
intervals are generally less than the heartbeat interval, and so waiting for the
next heartbeat effectively rounds them out.


Former-commit-id: eec58e7c8c86c4d44c3ab1763c7e5ce66adbb557