SERVER-16102 Adjust ReplCoordinatorImpl::canAcceptWritesForDatabase() to not require a mutex.
Because callers of canAcceptWritesForDatabase must already hold the global lock
in at least an intent mode, and because a node cannot transition between
write-accepting and not-write-accepting without holding the global lock in
MODE_S or MODE_X, we can arrange to set a flag indicating the answer to the
canAcceptWritesForDatabase question that may be consulted without taking the
coordinator's mutex or running in the executor.

This removes synchronization on the replication system from the write critical
path, and from the read path on primaries.  Future work will remove it from the
read critical path on secondaries.


Former-commit-id: c93641756a1d8444876b9b79c3aa9a9635626891