SERVER-16624 Do not use deadlock detection in invariant checking
The problem is that our deadlock detector does not check for deadlock
stability currently. It is possible that as the graph traversal
progresses, locks which we were waiting on earlier get freed and suddenly
become waiting on us, becoming counted in the cycle.

The fix would be to add check for deadlock stability by re-iterating the
discovered cycle. Until this is fixed, we cannot use the deadlock
deterministically for assertions, only for retries (which is what we
already do).


Former-commit-id: 92982e76c034f864f4499a0145ce7a75dc682812