SERVER-16773 Get rid of ScopedLock and 2 memory allocations
This change removes ScopedLock from the RAII lock objects' hierarchy. This
eliminates two memory allocations and two acquisitions of the Parallel
Batch Writer mutex.

I did not see any significant performance improvement, but next change
would be to remove the allocation of the PBR mutex as well and also to
make WriteBatchExecutor not allocate lock objects.


Former-commit-id: fe3e92d4257b30f01b62d4ef941686b7e0138a8c