SERVER-17312 collmod command now handles parsing of all arguments
For the two currently supported engine-specific options, the collmod command
will handle parsing the arguments and tell the CollectionCatalogEntry to
update it's flags option.

This removes the ability of storage engines to have custom options that can
be changed after the collection is created. There were issues related to
argument validation and replication of changes (including for initial sync).
A correct solution will be designed as SERVER-17320.


Former-commit-id: ae18bbec33db1c4bf57d328505474eb7bba517df