Optimize insertions into capped collections
When insertions need to delete the first key, they need to first find
the first key. Finding the first key is pretty expensive if the
beginning of your keyspace is filled with tombstones. This patch helps
because it caches the first key that is alive, so that we don't need to
go through all the tombstones.

It's the same optimization we already used for oplog. I just made it
more generic now.


Former-commit-id: 8afae158088de1034a5d897b677779377f6a70e8