SERVER-18350 Refactor expression_algo.cpp.
Added special handling for NaN values to be consistent with what
ComparisonMatchExpression::matchesSingleElement() does.

Added support for a point query to be answered by a partial index with
a range filter over a single field.

Changed to use static_cast instead of dynamic_cast because the
MatchExpression subclass is known based on the matchType().


Former-commit-id: eede63734af75cb50d5731e2677359fd8b067a0c