SERVER-19111 Move common addShard functionality to CatalogManager
This change is prerequisite for implementing the addShard method in the
replica set catalog manager.

It moves common addShard functionality as static methods on the
CatalogManager and adds the ability to create detached shard objects to be
used for making calls against shards and ability to retrieve the shard's
dynamic connection string.


Former-commit-id: 427b07f64530291af6ae410902939c58db9c0fa5