Don't lazily initialize _writeBatch
To properly implement read-your-own-writes, we need to create every
iterator with access to WriteBatch. That way, if a iterator was created
before the first write happened, it will be able to read that first
write because it will have access to _writeBatch.

This is consistent with mongo-rocks repository, as implemented in
https://github.com/mongodb-partners/mongo-rocks/commit/af0dedb94d2f13e7954af054db044905159ceeb0

Former-commit-id: f01a7735279f54bd86a8723079eab848f9bfad44