SERVER-20168 Add option to $unwind stage for treatment of nullish values
Adds a preserveNullAndEmptyArrays option to the $unwind stage. If this option
is specified, the $unwind stage will allow documents with nullish values for
the specified unwinding path, or those with an empty array as the value for
that path, to pass through unaltered.

Note this does not affect the behavior of non-nullish, non-array values.


Former-commit-id: 3cb6039e5d2968b1e361bf592452b95c7a76c770