SERVER-21522 Make a copy of the $config before running a workload.
Mutating the "threadCount" property in the ThreadManager is now safe
because subsequent executions of a workload group in the schedule will
use fresh $config objects. The proportion of workloads within the group
will be based off the number of threads initially requested, even if a
previous workload needed to be scaled down.


Former-commit-id: 721048cbc4e6b7d250cd729967b94e244426156d