SERVER-21749 Ensure $lookup obeys maxTimeMS limit
The $lookup stage uses a DBDirectClient to do its queries, which shares
the same OperationContext as the aggregation pipeline, but does not
share the CurOp, so does not inherit the time limit. To ensure a query
started this way obeys the maxTimeMS limit, a newly constructed CurOp
should inspect the OperationContext to calculate its remaining time.


Former-commit-id: 3191dbe15c70f1eb6546260d61fe9b0c7d10e045