Removed separate log backup and destination subdirectory logic.
In Hot Backup for the Fractal Tree Storage Engine, we do not need a
second directory in the destination directory.  This is because in 3.0
there is no user settable directory for the OpLog nor for the Fractal
Tree recovery log.  This fix only allows taking backups of the files
and directories in the current dbpath.


Former-commit-id: f261998f6665a19529ef92895f4cd25ecda83a37