SERVER-11064 Stricter validation of index key patterns.
Each value in the index key pattern must be one of the following:
  - a number > 0 (ascending)
  - a number < 0 (descending)
  - a string (special index type)

Since the key pattern is validated on startup for all existing indexes,
mongod will fail to start up if an index not meeting the above criteria
exists.

Additionally, if an index with an invalid key pattern is replicated from
an older version, then newer versions of mongod running in a
mixed-version replica set will fassert().


Former-commit-id: ffe490c80ecea03b02a0b20d55915c4efd2805e8