SERVER-22349 Throw exceptions from loadStored
The JS engine's loadStored eats exceptions that occur while it's loading
functions from system.js.  This also eats interruption exceptions, which
can lead to a situation where a map reduce job is killed during
loadStored, but the interrupt is lost. For tests where the map or reduce
stages are long or non-terminating, and we rely on killing them, this
can lead to hangs.

Removing the try/catch block around loadStored fixes this behavior.


Former-commit-id: dfc320fe9c8a5227b08c77a87f52996cf40b0206