SERVER-21881 Use fsync+lock when checking dbhash in concurrency suite.
The fsync+lock operation prevents the TTL monitor from running on the
primary after awaiting replication and before getting the dbhashes from
all the nodes in the replica set. Otherwise, a primary's dbhash can
reflect a delete operation that hasn't yet replicated to all of the
secondaries, leading to a spurious dbhash mismatch.


Former-commit-id: 41f296cdc37f92f8de0ea3d17109693c4f0befb6