SERVER-22691 change WiredTigerKVEngine init order
The KVEngine's _journalFlusher depends on the _sizeStorer and so should be
initialized afterwards to prevent a use-after-free error on shutdown. The mutex
for _identToDrop should also initialized beforehand rather than after.

In addition, shutdown calls to internal components in
WiredTigerKVEngine::cleanShutdown() should be ordered in a way that matches
their actual destruction order.

Former-commit-id: 95ca58f7396c1c4e1c5e54caa81dda875ca87b57