SERVER-22906: Cut down number of tcmalloc size classes
Cut down number of size classes from 52 to 42.
The fewer sizes classes we have, the less memory we waste in overhead with unused allocations,
but the more we waste in roundoff error. We use 12.5% as a threshold here, extending the
conceptual limit that already existed in the code to how we merge size classes.


Former-commit-id: 7be0eae950a33d49aaae967db250c265ccdfe463