Reinstated durability flag usage for fractal tree storage engine.
Because of recent KVEngine API changes, we needed to push the handling
of the durability flag into a lower layer of the fractal tree engine.
This durability flag does not change the durability of the actual
underlying write operations or logging, but does allow higher levels
to return correct errors, which in turn gets high level tests to pass.


Former-commit-id: a514812e7675c5b89470033ca06abef5a15f48a7