SERVER-22667 Move Chunk::pickSplitVector/MedianKey out of Chunk
The pickSplitVector and selectMedianKey utilities logically belong to the
balancer, so moving them out of the Chunk class and into shard_util.

The changes to balance.cpp have no functional effect, just rearranging
them to better match the balancer control flow and to make the subsequent
changes easier to follow.


Former-commit-id: 3c5ee8b721d3ea5b18d22527fcbea4e2e15e79ad