SERVER-23686 Handle race condition in ThreadPoolTaskExecutor shutdown logic.
Some implementations of NetworkInterface, particularly NetworkInterfaceMock,
cannot be shut down before the ThreadPool used by the ThreadPoolTaskExecutor,
leaving a race between shutting down the pool and scheduling the final completed
network operations into the pool. The workaround is to let the
ThreadPoolTaskExecutor's join() method execute any operation callbacks that get
left behind in this scenario. The safety argument is described in a comment in
the code.


Former-commit-id: 2b8ceb83a66eb400aa7dcad0a3735f2d6515b755