SERVER-23826 Run the dbtests under UBSAN
Also cleans up some handling around how undefined behavior
sanitizer finds the symbolizer, and suppresses some undefined
behavior in mozjs that prevented the dbtests from passing.

In addition, it was recognized that we can reduce the scope
where we inhibit UBSANs checks for misaligned access to only
wiredtiger, so do that too.


Former-commit-id: c60dea0181a621169f5130bd2fc291185421001a