SERVER-23814 Add sharding state initialization to secondaries
Outside of the OpObserver, there are currently 2 places where shard initialization occurs: during startup and during transition to primary.

This patch takes out the if !replSet condition in startup so secondaries will now also perform shard initialization. This also means that if the node will eventually become a primary, it will perform the initialization twice. However, since it is already initialized, most of it will be no-op the second time around.

In summary:
- secondaries initializes sharding state.
- nodes that will become primary will do what secondaries do. And on top of that, will read the minOpTime document and update it if minOpTimeUpdaters > 0.


Former-commit-id: 2423026ef06c43d7de4fa54095702698b359a4f9