SERVER-24104 sfinae duration ctor overloads
We can't just static_assert in the duration constructors, because we need overload resolution to
continue and find things like implicit conversions that then dispatch to default copy constructors.

This adds standard required sfinae to those ctors.  In other cases where the standard uses sfinae,
it's either not needed for us (because we don't require extension of duration types and we only have
integral duration reps) or appears not necessary now (several arithmetic overloads could
theoretically have other targets, but for the moment I think it's unlikely that we'd expect
Milliseconds{} / MySpecialNonRepType{} to actually work).  We'll have to revisit sfinae in those cases
if we ever do.


Former-commit-id: fb052c13742deab981c7ba729dad36e33bdd7393