SERVER-23114 Move functions involving dotted paths to separate library.
The ability to specify a dotted path (e.g. "a.b") to traverse through
embedded objects and array elements isn't defined in the BSON
specification and so it doesn't belong in our BSON library.

The following functions have been defined within a 'dotted_path_support'
namespace and accept an additional BSONObj as their first argument to
replace the associated method on the BSONObj class.

  - extractElementAtPath() is functionally equivalent to
    BSONObj::getFieldDotted().

  - extractElementAtPathOrArrayAlongPath() is functionally equivalent to
    BSONObj::getFieldDottedOrArray().

  - extractAllElementsAlongPath() is functionally equivalent to
    BSONObj::getFieldsDotted().

  - extractElementsBasedOnTemplate() is functionally equivalent to
    BSONObj::extractFields().

  - compareObjectsAccordingToSort() is functionally equivalent to
    BSONObj::woSortOrder().


Former-commit-id: cecbe424d32cbb475d9b0384d29b98a9fba9c89f