SERVER-24183 persist collation version string in catalog metadata
Also add an fassert() which trips if the icu::Collator attached to a
collection or index has a version string which does not match the
version string persisted in the catalog. This will ensure that the
icu::Collator always matches the collator used when the index was
initially constructed.

The collation's version string is simply the ICU release number. All
collations should currently be version 57.1.


Former-commit-id: 72b444d2831f018f3ae27219d07c69e6a3a60213