SERVER-24781 Temporarily prevent the balancer from taking collection lock
This change reverts to the old (pre-3.3) behaviour where the shard is
taking the collection lock on moveChunk command instead of the balancer.

It is done in order to stabilize the continuous stepdown suite and will be
reverted once the local dist lock manager is made to inherit all its locks
on restart.


Former-commit-id: cc904854866b1cd1a25508573eeab322c0dacbff