SERVER-25006 asio indicateUsed != indicateSuccess
Connection Pool ASIO connects indications of success with indications of use, which causes the
underlying connection pool to assume a connection is fresher than it may actually be (we said we did
IO, but actually we just haven't caused an error).

Because the connection pool it self checks connections out, marks them ok and then checks them back
in to trigger a refresh, this causes connections to always seem fresh, never be heartbeated and
never check against minConnections for dropping.

Change indicateSuccess to not also mean indicateUsed in conn pool asio and call indicateUsed where
needed manually in the one location where it's relevant.


Former-commit-id: e39a2a87eaca8a573b8b1145cf520415904019f5