Stabilize audit tests by fixing withinTheLastFewSeconds and surrounding
The problem was in that importing data from auditLog.json was done
with mongoimport starting in the separate process and it could
sometimes take more time than just couple of (milli)seconds.
With heavily loaded system, it could take up to 10 seconds, which
was breaking the following condition in audit collection query
that expected the entries to be within couple of seconds before
import was started, but that was not true for the cases when
mongoimport took long.


Former-commit-id: f3998894119516ee6c57f1318f166f27ac920bc4