SERVER-27304 add internalQueryAlwaysMergeOnPrimaryShard setParameter
When set to true on a mongos, aggregation queries routed to
that mongos which require a merging shard will always select
the primary shard as the merger. Set to false by default,
which means that the merging shard is chosen at random
amongst nodes participating in answering the aggregation
query.


Former-commit-id: 3f54fef03fdd7bd787a57387b84d298c9dd0222d