SERVER-27710 Move SNMP init hook out of db.cpp.
The `src/mongo/db.cpp` file is compiled into the main server program,
and causes broken-dependency issues in linking the enterprise modules.
By moving this symbol to another file, we can better express this
dependency.  Further, this change lays the groundwork to switch from
from a pointer to function to a `std::function< void () >` object.


Former-commit-id: 0d89f9f90e4f6a7ab00ce406b459d450b63e6ef3