SERVER-26990 Unify tracking of secondary state between replication and topology coordinators
This fixes a bug in the 499db1dc1834bce5725149ca7585152a390d3be9 [formerly 6adc71f6cf069803f9c1288aef88ffe0d21c6ffe] which caused crashes when a
sync source change was requested of a node not in the configuration.  It also fixes a dependency
problem affecting the shared library build.


Former-commit-id: c88c4809c2440d286ed0fc29e1e8d684f015e563