SERVER-29197 initial sync now builds capped indexes correctly
In initial sync, capped indexes were built along with regular indexes, as documents were inserted, using the MultiIndexBlock.
This structure requires that no records are deleted before committing the index builds.  But the collection cap might delete records prior to commit.
The new way to build capped indexes will be to create the indexes on a capped collection prior to inserting any records, for initial sync.


Former-commit-id: 3a3a6def395bab81e7545b15a93bee7799d9efb1