SERVER-29405 delay deleting orphaned shard chunks
When deleting the donor range after migrating a chunk off of a shard,
the range deleter will schedule the deletion at some time in the future,
according to a server parameter orphanCleanupDelaySecs, which defaults
to 900, or 15 minutes.  It does not delay range deletions preparatory to
migrating a range in, and does not put off deleting the donor range if
the moveChunk command has set the option _waitForDelete.

The file jstests/sharding/write_commands_sharding_state.js had CR (0x0D)
line endings, which made the patch fail lint.  The substantive changes
in the file were to add "_waitForDelete" options to the moveChunk
commands.


Former-commit-id: c63465a42ed89ee6563841d7b349fa85de69963e