SERVER-31440 Fix Connpool HostTimeout race
The executor connection pool host timeout is racy with respect to other
code that's unlocked the parent mutex to allow for callback execution.

While effort was spent to protect against background threads with active
requests and those participating in refresh, after those tasks have been
executed we race with callbacks in how quickly they can return.  When we
lose that race, we destroy the specific pool out from under those
callbacks.

Fix that by adding an ActiveClient wrapper that ensures a refcount on
the specific pool is increased for the lifetime of those calls.


Former-commit-id: c3e174cab7b8e4a19772746942c7e68daa53bc5e