SERVER-31965 Correctly handle certificates for SRV URIs
The hostname provided by SRV records is a canonicalized FQDN ending
in a '.' character.  X.509 certificates use a canonical hostname
with the trailing '.' removed.  The comparison between these two
forms needs to strip all trailing '.' characters.  This is
considered safe in all cases, as a DNS spoofing attack would still
require forging or obtaining a certificate with a canonicalized name
to make a redirection work.


Former-commit-id: c2d309d23cf918e1ded8fc241a1c2108dd0e31d3