SERVER-31133 Ignore mechanism for `StatusWith`
Change `StatusWith< T >::getStatus` to return `const Status &&`, which
permits an overload for `Status::ignore() const &&`.  This would be
called from expressions which use `StatusWith< T >::getStatus()`
for example, `f().getStatus().ignore()`, giving a straightforward
and concicse way to indicate that the success or failure of an operation
is irrelevant (and thus the result is also irrelevant).


Former-commit-id: e2733b633fbd06b380c4f0febca3e2f33e93dba5