SERVER-33437 Make Decorables not type-erased.
The current Decorable system type-erases the most derived (Decorated)
type when registering Decorations.  This is illegal as derived classes
which are no longer standard-layout can have the `Decorable` base
class reordered with respect to the base address of their storage.

This also removes the "with-owner" forms of declaring a decoration,
which caused problems with layout.  The decoration handles have
been modified to provide mechanisms to get back to the owner
directly, given a pointer or reference to the decoration.


Former-commit-id: 2084f6702cd026bebe61818564814aac3e2f12a3