SERVER-27534 All writing operations must fail if the term changes
Note that in our master branch, we committed a more ambitious fix for
this, which modifies lock acquisitions to always check for interrupt
(obviating the need for the interrupt checks in these batch modify
functions). That will hopefully prevent future problems like this.

On the 3.6 branch, though, we just fix the case described in the Jira
ticket by moving interrupt checks to happen after we take a collection
lock (which helps because a stepdown can't occur while the lock is
held).


Former-commit-id: 97f6f0e4612fa04f7268c6f354c7065281075267