SERVER-31128 Skip validate on NotMaster error response.
This change is in preparation for running the validate command against
secondaries in a replica set.

The fuzzer test suites require that the validate command is only run
when the server's featureCompatibilityVersion=3.4. Otherwise, it is
possible that previously inserted Decimal128 data could lead to an
actually innocuous data consistency failure.


Former-commit-id: 2c5a4238a190f52dd2a301524c751aa3fac3a8d8