SERVER-35514 Make max_time_ms.js even more robust.
There are rare cases where interrupting javascript execution on the
server with a stepdown or timeout causes an InternalError, which
causes some test commands that were expected to fail with
ExceededTimeLimit to instead fail with InternalError. For now, we just
accept InternalError as an acceptable failure when the command
executes java on the server (i.e., uses $where).

This change also includes the error in the failure message for some
assertions, which may help with diagnosis of other problems in the
future.


Former-commit-id: 361ec85b0a73baac9033cf5fe646e1cb40dc8f26