SERVER-36676 spinlock as mutex in dbg builds linux
We have some lovely deadlock detection code built into the hang analyzer
on linux which does a great job with our internal lock types as well as
with mutexes. For debug builds, backing SpinLock's with mutexes gives us
that deadlock detection for free.

This constrains them from being released on a different thread, but
critical sections already have that property, so we can't have been
relying on it.


Former-commit-id: db1535d7e9094fd444609ade9bf0fb725d611f85