SERVER-36859 Validate dbname in DurableViewCatalogImpl::iterate().
If the DatabaseImpl constructor (which calls
DurableViewCatalogImpl::iterate()) throws an exception, the partially
constructed DatabaseImpl trips an invariant as it gets destroyed
during exception unwinding. This change makes sure that
DurableViewCatalogImpl::iterate() properly validates the 'viewOn'
namespace, so that it doesn't throw an exception when getting parsed
as a NamespaceString.

Former-commit-id: ec132343b7e7225e36575f173c242b383f505eb2