SERVER-37358 dont early return from ARS
As part of SERVER-35689, Replica set targeting was made async (but with
blocking in the ARS).  But when we block on the returned future, we do
so via an opctx which isn't marked uninterruptible.

This means that opctx killing, or timing out, during scheduling leaves
dangling references to the ARS and the baton in the ARS destructor.
This in turn causes the invariant inside the baton to trigger.

The fix for now is to make that an uninterruptible wait


Former-commit-id: b757d87fc622d44b16013e2722832a655d7ca052