SERVER-38754 Make speculative majority change stream reads wait on optimized optimes
Speculative majority change stream oplog reads now only wait on the latest scanned oplog timestamp when possible, as opposed to always waiting on the latest system-wide lastApplied optime. If a document post-image lookup occurs locally for a change stream read, then this optimization is not safe, since the document lookup may reflect data at some unknown timestamp. In this case, we revert to waiting on the node's lastApplied optime.


Former-commit-id: ffb19c17927a5a0a07a7cd7cd7744cea80ab3e26