SERVER-37560 Enforce monotonicity of speculative read optimes automatically and rename setSpeculativeReadOpTime
Allow clients to set speculative read optimes on SpeculativeMajorityReadInfo in an order that is not monotonically increasing. Instead, when a new optime is reported, we only update the current speculative read optime if it is less than the new optime. This makes it easier for clients to utilize this class in the cases where they may not know what other clients are setting read optimes and when they do so in the course of a command execution.


Former-commit-id: 36e9a68390989b87718b8df3a7d419ba870f9692