SERVER-39000 Fix Unittest Framework initialization.
Some tests in `dbtest` do not use the `mongo::unittest::Test` class
as a basis for their implementation and thus need some way to benefit
from global DB Environment initialization functions.  Specifically
the changes in SERVER-32630 required `serverCompatibilityVersion`
to be set sensibly.  Some tests in `dbtest` were not correctly
getting this benefit; instead only incidentally getting a correct
setting by accident, as the results of an unintended residue of
an earlier operation.  This can lead to inconsistentcies in which
tests pass, as link order changes -- the tests are registered using
static initialization, whose instability of order can cause
mysterious failures in `dbtest`.


Former-commit-id: 677a65ba43fecf7cf20b7a2c34b5ec2bed413edd