SERVER-39693 Add unit tests to pcrecpp
Note: Originally this patch also set up pcrecpp_unittest.cc to run
as a C++ unit test along with the rest of the server. This caused
some issues, so we abandoned it. I am leaving the new unit tests, however,
in case we need to patch PCRE again and run the tests during development.

Changes to PCRE made by MongoDB (not the original authors) are tested
through the $regex expression.


Former-commit-id: b3bd75fc2f404931729e0c00992053b03859f817