SERVER-39380 Update vendored scons to fix --config=force behavior
When SCons is run --config=force it overwrites the Decider for the
environment passed in to a decider which will rebuild everything. This
occurs even if you do no retrieve configures environment via
env=Conf.Finish(). Now when --config=force the configure logic will
clone the provided Environment, change that environment's decider, and
if you retrieve the environment via Conf.Finish() the logic will reset
the decider to the one set in the env passed to Configure().

This is a cherry-pick of the commits for
https://github.com/SCons/scons/issues/3303 into the MongoDB vendored
SCons tree.

Closes #1295

Signed-off-by: Andrew Morrow <acm@mongodb.com>

Former-commit-id: 2ecd83f822b242db1c13f2abc3ff3593ba622149