SERVER-37179 Pull out starting transaction from session checkout and push it down to before command execution.
This patch redid f4b408bbeb [formerly 248601a647] and 45628417dc [formerly 4fb38d9c10] from master on v4.0 branch.

Transaction will begin or continue after waiting for read concern. If
an error is thrown on starting transaction, it'll be able to wait for
write concern if a write concern is specified.


Former-commit-id: 5366d3c6ea014f1bd19eee1a149f46f3b1227a2b