SERVER-40158 Make failure in ::eventfd non-fatal
After introducing the linux baton more widely (on each client, on
creation, rather than during an ARS dispatch), users are more likely to
encounter that condition, and we can handle it more gracefully.

By changing the failure to create an eventfd into a throwing exception,
we'll cause new client object creation to throw, which will cause us to
close the incoming client socket.


Former-commit-id: 4414263139990f34e93eceebfc4ec2ead2ff4373