SERVER-40165 set wtimeout deadlines with the precise clock
In ReplicationCoordinatorImpl::_awaitReplication_inlock, we generate our
deadline for waiting on wtimeout with the fast clock. Because we
generate the deadline with the fast clock, but actually wait on it with
the precise clock, it is possible for us to wtimeout early if there is
significant drift between the precise and fast clocks.

This is unlikely to happen very often in practice, but could happen if
the thread responsible for pushing ahead the fast clock get's
descheduled for a substantial period of time. In that case, it is
possible for the fast clock to drift multiple seconds behind the precise
clock, and to immediately timeout in
waitForConditionOrInterruptNoAssertUntil.

(cherry picked from commit 1967db1f68fba4ba22d1d60b0ce602564562fc1d [formerly 46086585d49f2da53d46c3f121e9f8dc23b699a1])


Former-commit-id: c2da4180540e81ba591b858e14688d40f9455554