SERVER-41056 Fix NiTL::shutdown() without startup()
It should be legal to call shutdown() on a network interface without
calling startup() (because an owning task executor will call shutdown()
in its dtor).

Currently this crashes (because the NiTL doesn't keep track of whether
it's started). Changing the _inShutdown atomic to a _state enum atomic
can resolve that problem


Former-commit-id: ba825c452525fb6fa7e28196cc84a49809b6f4be