SERVER-41766 Update the on-disk multikey flag in a side transaction when running inside a multi-document transaction
When a write inside a multi-document transaction needs to set an index as multikey, we update the multikey flag in the on-disk catalog in a transaction separate from the parent transaction. We commit this side transaction immediately, so as to avoid the catalog write generating prepare conflicts if it was written as part of a parent transaction that later became prepared. In general, it is safe to set an index as multikey too early. The multikey write is timestamped at the most recent value of the LogicalClock.

(cherry picked from commit 7f37962a1f7e7b2059e701c76f3022878a310c34 [formerly 3147f5e1c37546b817934ef892d5e353170a9935])


Former-commit-id: e549b451eb4ba9aa11f79a2356061f5dab6c943b