SERVER-42518 Prevent paths with successive array indexes from using a wildcard index.
Consider a query on a generic path "c1.c2...arr.0.1...cn"
consisting of n path components. Somewhere in the middle of
this path, there is a component "arr" that is known to be
multikey, followed by multiple successive array indexes ("0"
and "1" in this example). Any predicate whose path matches
this description cannot correctly be answered with a
wildcard index. This is because wildcard indexes do not
recursively expand directly nested arrays. That is, when
indexing an array such as [[1, 2], [3, 4]], the index keys
will be [1, 2] and [3, 4], as opposed to 1, 2, 3, and 4.
Since the index does not contain keys for individual nested
array elements, it cannot be cheaply used for queries
against nested array elements.

This change fixes the planner to avoid attempting to use a
wildcard index in such cases, thereby fixing a query
correctness bug.


Former-commit-id: 0f0463d0442b2c77cb650a8e84fea4d634fb63e6