SERVER-41357 SERVER-41360 SERVER-43324 Add Mutex, BasicLockableAdaptor, and DiagnosticInfo
This commit backports Mutex, BasicLockableAdaptor, and DiagnosticInfo to
v4.2. These pieces came in completely in various interleaved commits.
The constituent commits partially included in this backport are listed
below. The majority of these commits are part of the "Improved
Diagnostics for Latches" epic.

Mutex:
- SERVER-41357
- SERVER-41362
- SERVER-42893
- SERVER-42595
- SERVER-42165
- SERVER-42895

BasicLockableAdaptor:
- SERVER-43324
- SERVER-43374
- SERVER-43800

DiagnosticInfo:
- SERVER-41360
- SERVER-41362
- SERVER-41358
- SERVER-41364
- SERVER-42448
- SERVER-42363
- SERVER-42363
- SERVER-42492
- SERVER-42595
- SERVER-44086

Note that while this commit adds these pieces and integrates them into
the mongo-server codebase, it does not convert existing stdx::mutex
construction. It also does not include the updates to Interruptible that
were done as part of "Improved Diagnostics for Latches".


Former-commit-id: 3a05e4ce75cd03b09b52ec5943e1801c6d91279b