SERVER-44823 Add sharding support for $planCacheStats.
When a $planCacheStats pipeline is delivered to a mongos, it
will be forwarded to one host in each shard following the
normal host targeting rules for aggregate operations. Each
of the documents returned to the client will be augmented
with "shard" and "host" fields, giving the name of the shard
and the host:port string from which the plan cache metadata
originated.

Allowing clients to collect plan cache information from
every shardsvr in the cluster is left as future work.


Former-commit-id: 01e9fd6d0ec033089ba2f45a440619d3a08c9d9f