SERVER-43721 Make the AuthorizationManager use DistCache
The DistCache (to be later renamed to ReadThroughCache) was derived from
the same implementation under AuthorizationManager and this change
removes the code duplication.

In addition, it makes the following changes to InvalidatingLRUCache and
the DistCache:

* Simplifies and optimises the InvalidatingLRUCache:

The way it is implemented now, it performs up to 3 operations per
lookup, unvalidates entries unnecessarily and has overly complicated
logic, which is source of a crash. Instead of fixing the bug, this
change rewrites it in a simpler way, which introduces a ValueHandle
instead of bare shared_ptr for the return value, and only performs
additional work if entries fall off the underlying LRUCache.

* Moves the DistCache under src/util and adds unit tests:

This change pulls the DistCache (which is the main consumer of
InvalidatingLRUCache) into its own library and moves it to be
under src/util like the other caches and adds unit tests.

 delete mode 100644 jstests/auth/pinned_users.js
 create mode 100644 jstests/auth/pinned_users_clear_pinned_user_list.js
 create mode 100644 jstests/auth/pinned_users_exclusive_lock_on_admin.js
 create mode 100644 jstests/auth/pinned_users_remove_user_document_unpins_user.js
 create mode 100644 src/mongo/util/dist_cache.cpp
 rename src/mongo/{db => util}/dist_cache.h (56%)
 create mode 100644 src/mongo/util/dist_cache_test.cpp


Former-commit-id: 73b89c6fc4ed6279b52e2588c102c7fc1182189b