SERVER-40620 uassert and log when fetching dangling index entry
If a FETCH_STAGE encounters a record id that does not reference any
existing documents, that means either the document was deleted since
query execution encountered the index entry or their is corruption
somewhere. If the snapshot id indicates that the query has not yielded
since the time that the index entry was loaded, that leaves corruption
as the only possibility. We return and error and write to the log with
instructions on how to address potentially inconsistent data.

 create mode 100644 jstests/noPassthrough/query_yields_catch_index_corruption.js


Former-commit-id: 8e810c9ff48db40e63b3201f8bd4d0bf4fdea8ce