SERVER-45261 create more state in NiTL ctor
Calls to NetworkInterfaceTL::startCommand() before
NetworkInterfaceTL::startup() wil cause crashes (due to an assumption
inside NetworkInterfaceTL::startCommand() that we have a connection
pool.

That's out of contract for a NetworkInterface

To fix that, lets create more state on the NiTL in it's ctor, and only
spin up the background thread in startup().


Former-commit-id: 1878d44a2ac3764d71f00e8ce04ecb9911ab6023