SERVER-45117 Guard NetworkInterfaceTL::setAlarm() more aggressively
The underlying issue is that NetworkInterfaceTL::_state and
NetworkInterfaceTL::_inProgressMutex do not synchronize with each other.
This is at best a temporary fix. Reasoning about alarm ordering in the
context of the NetworkInterface itself is messy. This approximates what
a proper composible AlarmScheudler type looks like. Hopefully, there
will be more consistent work under SERVER-46419.

Former-commit-id: 57f6385025adf630a410f9d658f61b5afd140121