SERVER-45765 Race in initiateWithAnyNodeAsPrimary
If there's an election between the call to replSetInitiate and the last call to
replSetReconfig, the test would nevertheless expect Node 0 to be primary and
all others to be secondaries. Fix it so will continue as soon as all nodes are
primary, secondary, or arbiter.

Also add resource_intensive tag to a test that starts 12 nodes.


Former-commit-id: 38ec223f7478be14fc3bf082643a1109efaeb57c