SERVER-37637 Standalone nodes should not restart in-progress two-phase index builds on startup
If a replica set node is started in standalone mode and there are
incomplete two-phase index builds, instead of restarting them, we mark
the indexes as 'frozen'. This signifies that the indexes are unusable,
and will not be rebuilt. The only option available to users is to drop
the index. This can corrupt the replica set, but modifying replicated
data in standalone mode already poses this danger.


Former-commit-id: df03d4ddc092ca3616575dd6ec08f7c4852d51e0