SERVER-46008 Remove thread specific attributes feature from boost::log
Because we prefer to compile with BOOST_LOG_USE_COMPILER_TLS for performance reasons it is not safe to log during global shutdown because of the uses of thread_local.
We are not using this feature so we remove it instead of removing BOOST_LOG_USE_COMPILER_TLS and affecting the server.

The issue only affects unittests as we only do clean shutdown there.


Former-commit-id: 23a5160098996b401a1145504a7be356a3e6c8f1