SERVER-46894 Wait for the current config to be committed before running reconfig.
This also changes the behavior of waiting for both Config Replication and
Oplog Commitment proactively at the end of safe reconfig. Both conditions guarantee
the safety of the next reconfig. Now we only wait for Config Replication at the
end of safe reconfig and wait for both condition at the beginning of the
next reconfig when the waiting is necessary.


Former-commit-id: 89ec7322a58686b89aa71f26b1f050ded94cf949