SERVER-47425: 4.2 binaries starting up on WT log version 4 should continue to write log version 4.
In the upgrade/downgrade lifetime of a MongoDB dbpath, WT log version 4 is first
introduced when a 4.4 binary downgrades in FCV 4.2. Once a dbpath is touched by
a 4.4 binary, the datafiles left behind cannot ever be interpreted by a binary
version earlier than 4.2.6+.

In this patch, 4.2.6+ will discover whether it's starting up on log version 3
(left behind by 4.0 and 4.2.5-) or log version 4. When starting up on log
version 3, it will continue to write log version 3. When starting up on log
version 4, it will continue to write log version 4. There will no longer be
a path to writing log version 3 without performing a resync.


Former-commit-id: 77036e27c1343fde8cabd5285acbf09b0c69b69e